+++ kitty/options/parse.py
@@ -1185,6 +1185,9 @@ class Parser:
     def shell_integration(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['shell_integration'] = shell_integration(val)
 
+    def show_hyperlink_targets(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
+        ans['show_hyperlink_targets'] = to_bool(val)
+
     def single_window_margin_width(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['single_window_margin_width'] = optional_edge_width(val)