+++ kitty/state.c
@@ -335,6 +335,8 @@ destroy_window(Window *w) {
     Py_CLEAR(w->render_data.screen); Py_CLEAR(w->title);
     Py_CLEAR(w->title_bar_data.last_drawn_title_object_id);
     free(w->title_bar_data.buf); w->title_bar_data.buf = NULL;
+    Py_CLEAR(w->url_target_bar_data.last_drawn_title_object_id);
+    free(w->url_target_bar_data.buf); w->url_target_bar_data.buf = NULL;
     release_gpu_resources_for_window(w);
     if (w->window_logo.id) {
         decref_window_logo(global_state.all_window_logos, w->window_logo.id);